Little core changes cd

+++ src/vizdoom/src/viz_defines.h
@@ -28,7 +28,7 @@
 #define VIZ_VERSION 109
 #define VIZ_VERSION_STR "1.1.0pre"
 
-#define VIZ_TIME (unsigned int)(level.starttime + level.maptime)
+#define VIZ_TIME ((unsigned int)level.starttime + (unsigned int)level.maptime)
 #define VIZ_PLAYER players[consoleplayer]
 
 #define VIZ_DEBUG_PRINT if(*viz_debug) Printf
